Expand prettier coverage to root, format everything
Mostly this is a DevEx thing, where I would save a file and VSCode would
autoformat everything (with prettier extension installed and configured
to look for a prettier config). This is probably not completely atypical
so might as well set up the ignore appropriately and then format
everything that hasn't been formatted yet.